Network abstraction for supporting multiple network interfaces types
Re-arrange the network setup code base to allow support for multiple network
interface types and an easy integration of new interface types.
